Done for the day. I got quite a bit done.
Tomorrow I will deal with Cuda arrays.

I am thinking, instead of building Cuda constants in like this, and then doing the same for each and every type, I should use Spiral's powerful intensional polymorphism aspects to build those constants and functions directly into the langauge.

It can be done, that much is certain.

It is really a matter of where I want to put them, but there is definitely a level even beyond this.

I haven't really reached the pinnacle of language design. More can be done.

There are different ways of doing interop, but let me go at it slowly, one step at a time.

But the point I am now is already otherworly by any reasonable measure. Right now, the backend is in fact usable.

There is nothing stopping me from starting work on the ML library.

Nothing.

Right now the entire language is 3.7k lines of code. Honestly, I really feel like it could stand to go down at this point and the various aspects of should be slowly unloaded down onto library.

...No actually, doing it like this is fine. It will make it easier to switch backends, while doing the code generation affecting functions would lock me into a language.

Everything is perfectly fine.

...

But on the other hand, this kind of approach is something that would be unique to Spiral. Other languages definitely can't do it.

I'll keep it in mind.

Let me just finish this by saying that by the standards of any other language 3.7k lines is really nothing.

If I wanted a minimalist implementation I can do with far less, but if I want an useful implementation, I think I am going to be adding stuff either way.

Though I really can't imagine it growing to the dizzying heights that some of the other languages have which tells me that I am on the right path.

This is true metaprogramming without a doubt.