Right now I am reading the AMA by Silver and the other guy. I missed it a few days ago so I am catching up.
I am still thinking of what I should do next.

Well for starters, how about I aim to make a Cuda map kernel?

After that I will take a break from the backend work to finally put all the pieces together with regards to layouts and do the renaming that I wanted.

That seems like a plan.

I think that once I perfect the layouts, the Spiral language will be complete from a semantic standpoint. It already is, but I need to polish the Layout feature before I can call it that. Given that it is still named Record, I am not there yet.

This redesigning is something that I've done a lot during the devepment of Spiral and it has improved the language greatly.