Fuuu, I finally got this piece of code done. Now let me test it.
To be honest, I had not expected I would be doing this today. This kind of explains why the language is taking so damn long. I really should go an extra step and find some way of dealing with namespaces, but fuck that.

I still have to test this. Let me get on with it.