It works now. Good. I am adding the replacing of assemblies with Map to the list.
*1) Rename Record to Layout.
*2) Add Layout annotations for tuples.
*3) Make sure that the layout annotation is *stackable*.
*4) Add PackedStack layout which corresponds to sequential fully packed struct layout.
*5) In the Ty definition, merge the HeapFunT and StackFunT and all the types that will be introduced into a single case.
*6) Extend CudaGlobal so it takes a Ty.
*7) Extend Layout with the LayoutFsharp type.
*8) Make sure that layout can be set on types with any number of free variables.
9) Add the Sizeof op to the language.
10) Modify .NET calls so that when arguments are all literals and the return type is a primitive, they are called at compile time.
11) Add the .NET overrides for String and Array types.
12) Put in the conversion from closures to .NET F# functions.
13) Add conversions for F# records and mutable records.
14) Make the Extern module.
15) Replace the Assembly with a Dictionary.

Well, regardless of how today turned out and even if there was a lot of churn and the end result is hardly satisfactory, I did improve the ergonomics of the language so I can be proud of that.

I just remembered to rename PatType to PatTypeEq. That should help readability a bit.

I think this is it for the day for me. I focus hard throughout the entirety of it. Hopefully tomorrow I will be able to start work on Cuda arrays. I am still thinking through how to do Layouts.

Actually, before I start work on the arrays, I think I should deal with that.

Yeah, I should deal with that. It won't take me too long, maybe I'll manage it all tomorrow.

The really difficult parts of them I already did ages ago. The number of points is not really the big deal here, individually they are peeled down to the minimum anyway.

Let me just take a moment to say this.

The really astounding thing about all this .NET interop trouble is how shitty the API is. It is like it is made with the design patterns from 15 years ago, in the dark ages where things like tuples had not existed yet. Which is in fact the case.

F# really makes it looks so easy, but the quality of its interop is owed less to .NET as a platform that to the all the effort that went into the language. I would not bother with it.

Kudos to the F# team for this.