I wanted to make the commit earlier, but I delayed it in order to avoid rediffing the cache until I was done testing.
Today, the things are going really, really well. I completely revitalized both the codegen and the destructure function. Now they are much more readable and make more sense. Not to mention, with the changed semantics, I won't have to worry about printing the operator names for structs.

Those complicated template functions have been rewritten for full redability.

In hindsight I really did a shoddy job before, but better ideas simply did not occur to me back then.

The thing is, when you come back to the code you written after a few months and cannot understand what it does anymore, that is a sure signal that a rewrite is in order. And as it turned out, this time things are far clearer.

I think the way things are going, I really will finish the language by the end of the month.

I remember writting that the records, now layouts are the last big feature of Spiral and that is true. I just need to perfect it.

Let me remove StackT and HeapT and I will add LayoutT.