~*1) Rename Record to Layout.~ ~*2) Add Layout annotations for tuples.~ ~*3) Make sure that the layout annotation is *stackable*.~ *4) Add PackedStack layout which corresponds to sequential fully packed struct layout. ~*5) In the Ty definition, merge the HeapFunT and StackFunT and all the types that will be introduced into a single case.~ *6) Extend CudaGlobal so it takes a Ty. *7) Extend Layout with the LayoutFsharp type. ~*8) Make sure that layout can be set on types with any number of free variables.~ 9) Add the Sizeof op to the language. 10) Modify .NET calls so that when arguments are all literals and the return type is a primitive, they are called at compile time. 11) Add the .NET overrides for String and Array types. 12) Put in the conversion from closures to .NET F# functions. 13) Add conversions for F# records and mutable records. 14) Make the Extern module. 15) Replace the Assembly with a Dictionary. 16) Add union types to Cuda
Apart from #7 and #4 which I won't bother with right now, I am basically done with this.

1) Add StackPacked layout which corresponds to sequential fully packed struct layout.
2) Extend Layout with the LayoutFsharp and HeapMutable types.
3) Extend CudaGlobal so it takes a Ty.
4) Add the SizeOf op to the language.
5) Modify .NET calls so that when arguments are all literals and the return type is a primitive, they are called at compile time.
6) Add the .NET overrides for String and Array types.
7) Put in the conversion from closures to .NET F# functions.
8) Add conversions for F# records and mutable records.
9) Make the Extern module.
10) Replace the Assembly with a Dictionary.
11) Add union types to the Cuda side.
12) Fix the codegen for the Cuda side.

Tomorrow I will deal with adding conversions and extra layout types to the language. It won't be hard. I also need to go through the Cuda side, today I did not have time.

Today, I really did a brilliant job in various respects and significantly improved the readability of the compiler.

Tomorrow, I will deal with the other half of it on the above list, and probably won't bother with the rest for the time being. They aren't that important.

The important thing is that after tomorrow, the language will be pretty much complete and I will start work on the ML library in earnest.

Lately, I can pretty much track the improvement in my mental state the closer I get to the completion of this.

There is more from where that came from.

The worst part of the day for me is to stop work. If I could do it 24h I would, at least until I am done with my goal.

But unfortunately my brain and body need rest, so I will take off for now.