Hell of a night, as expected. But I am doing better now.
1) Make the Extern module.
2) Replace the Assembly with a Map module.
3) Extend term casting with `→`.
4) Rewrite pattern matching to remove the dependency on CSE. It should not be used to pass information like it does.

Let me start be redoing pattern matching. I've spent really a lot of time thinking how to do Extern.

This situation I am in is really a blessing in disguise as it forces me to deal with interop properly. And I will deal with it properly.

Once I get the above points out of the way, I will not just have interop with .NET, but pretty much every language imaginable.

In programming, you do not fight with your effort, but with your desire and imagination.

https://julialang.org/blog/2017/03/cudanative
https://devblogs.nvidia.com/parallelforall/gpu-computing-julia-programming-language/

Let me just say that it is too bad Julia did not have this stuff a year ago before I started this. The GPU stuff that it offers is impressive, but I am still skeptical of it.