I had some extra time and will in the last half hour so I decided to do some more programming. As luck would have it, I fixed the bugs completely and now the assemblies are being converted into Spiral's native modules directly on open.
This is brilliant. If I could do the same for standard Types, I will have complete .NET interop.

1) Make the Extern module.
2) Extend term casting with `→`.

Only two things left. #2 has already been partially started on in my work on doing interop with events.

Everything that comes after this will be a part of work on the Extern module so I do not need to belabor the point.

Since the last update for the Simulacrum blog there have been exactly 4 months. If I have nothing to show by the end of next two months it will be an embarassment for me.

Yeah, it has been a good long while, In terms of ML development I am not even at the point I was a year ago.

Now that I have the Cuda backend, it won't be a big deal to catch up.

There is still the small big matter of getting the interop done right.

As this recent exercise of translating the Assembly to modules shows, I've been holding back way too much. More, much more can be done.

At the end of it all, it will be unfortunate that I won't be able to draw out Spiral's full power on .NET playform - to do that I would need a VM with support for multi-staged compilation, but this will definitely be enough.

As a language, Spiral's power is trully awe inspiring. Without a doubt, it will show me the truth of programming.

And when my desire fractures and manifests into agents, I will be able to chisel it down all the way to the root.

And most importantly, along the way I will have a lot of fun!