I am done moving stuff to core, but I have not tested it yet. This was rather time consuming.
Now that I have the unbuilt ops at my disposal, it will be trivial to do module_add and module_remove.

But I will keep that for later. It is not important.

What is important is continually improving the language ergonomics.

I'll have breakfast and then move to testing this.

No doubt the recent changes will have broken everything.