```         p core <| fun core ->             loop module_auxes <| fun auxes ->                 p module_main <| fun main ->                     ret (module_open core (auxes main)) ```
Without a doubt this is the correct way to do this in Spiral. I guess now I know how to do the equivalent of unhigienic macros in the language.

It is amazing that just figuring out where to put Core took me half the time needed to do the entire rewrite.

Let me finally do the testing.