So parsing is responsible for about 1s and loading those extra assemblies is another 0.7s. That means that 4.3s is due to moving core_functions into Core.
Ok...

There is not much I can do about this.

It is good enough.

As much as it anguishes me that I do not have the proper tools to optimize the compiler itself.

It is annoying, it is so annoying and frustrating. I am looking at the future and seeing the big picture and this is it:

Just like the partial evaluator that is Spiral can be made to optimize a program on the lower level, so can the pevaller itself be optimized by fusing it with the libraries.

Nested partial evaluation is essentially functional composition applied to language compilers.

What this would do is not yield better programs, I think what Spiral gives me here is the pinacle of performance already, but what this would allow one to do is drastically optimize compilation speeds.

Also, today I had a lightbulb moment when I realized the implication of how a seemingly trivial move such as this took me the entire day.

Programming does not scale.

This is the hard cold truth.

I think at 4k LOCs now, Spiral is near the top of what I can manage with it. Probably if it exceeds 10k lines, it will become far too unwieldy to manage. I will have to add yet more stuff, but discounting the Futhark backend that I will eventually do, I do not think the language should exceed 5k LOCs.

In my last update, I think it was something like 2k LOCs, so it got rather fatter in the last 4 months.

In fact, the IDE is already getting sluggish with it.

With current technology, I think this is the best that is really possible, and if made by humans, nothing really revolutionary in this respect will come about.

Similarly, I think Spiral is really the closest a language will get to mainstreaming dependent typing.

I said here that programming does not scale. If that is true for the regular kind of programming I am doing here, doing dependent formal proofs would just raise the difficutly by 100x.

When I tried Coq and crashed, that was not a fluke. Coq is really the best dependent typing has to offer when it comes to doing proofs, and the fact that I had to give up over some stupidly trivial programs is proof enough that the gulf between vanilla programming and formal proving is trully enormous.

Idris won't be able to overcome it. Not a chance.

Hence the type propagation I am using here is the true path.

In terms of programming language development, we are approaching an evolutionary bottleneck.

The way nature likes to work is to flood the world with cheap organisms and then let loose the predators.

So slow languages (Python,Ruby), dull languages (C offshots like C++ and perhaps even Rust and D), complex languages (Idris), and the insufficient ones (F# and the MLs) are due for a cleansing.

The breaking point is the GPU.

A language of the future will absolutely have to support those things, produce high quaility code and high levels of abstraction.

It will essentially have to break the old trend of highly expressive languages being slow.

As much as Julia annoys me being a dynamic language, it is making all the right noises. It is not special as a language; other languages do specialization and type inference too, but Julia is the only dynamic language which actually makes an effort to focus on that.

It does not go a step further like Spiral and offer outright guarantees and it lacks great things like pattern matching and union types, but it might have a future.

Not much of one, after Spiral enters the picture. As a language, Julia is significantly complex. This is generally what happens with dynamic languages.

Programming does not scale.

I made the right choice by making Spiral. The edge it will give me, if I manage to achieve my goal in the end will prove vital.

But I must never forget the fact that it does not scale. I will run into limits with my programming skill like I have in all other endevors.

Spiral is a little bit of sunshine and little bit of hope and light.

It is my only chance to win this.

I must never forget that my victory conditions is reaching the next level of abstraction and letting the agents themselves create their own language with Spiral as the basis. No doubt, waiting for the damn thing to run will be torture for them.

If I tried to bring my ideas about what Spiral should be to fruition myself, it would take me a decade. I can't afford it. This design is essentially the local optimum for me.

All that I can do past here is to continually polish it.

I've gotten lost in thought. The day is nearly over, but I do not want to stop just yet.

Before that, let me at least plan out my next steps.