"The first thing I need to do in order to put in Extern Ops is is redesign how recursive types are boxed. I did a nice bit of thinking during the night, so I know how to do it, but there are some bigger problems with my approach.
The main one would be separating type instances from bare types. This is a massive pain in the ass.

I can't really have instances and apply them in a OO manner.

Instead they would have to be passed as an argument. The bigger problem is how complex .NET is. Having recursive types would make compiling to them difficult and then there is that annoyance in dealing with instances.

Then there is the fact that types might be extended in different assemblies.

There is also a difficulty that if I have both the Extern and the standard way of doing interop the way I do now, the two systems would be incompatible.

So for .NET, unless I need to use F# maps or something like that, I will keep the current system and fix the fact that it is not checking the staticness properly right now.

That having said, I will also bring in the changes to how recursive types are done as I need it for Extern. There is also the problem that having multiple join points in the same memoized methods dict might cause them to get overwritten in pathological cases."

I am almost ready to start. Whatever the case is, I need Extern, but it won't be enough for full interop. In fact compiling to it given the vagaries of .NET will be a pain.

...No, actually, I might be able to compile to it. At any rate, the ideas are still whizzing in my head, I haven't converged on the final design yet, that is for sure.

Just now as I am writting this, I am getting an idea of how to deal with the instance/bare type separation problem.

Well, let me think about it for a bit more and then I will start.