I did the changes that I wanted regarding types. Actually, I think that what I had previously would have worked to, but it was so convoluted that it simply was not obvious to me.
What I will do is change the way type is parsed back to the standard syntax. This one is doing me no favors.

...No actually, what I will do is just remove those `type_create`s...

The join pointers were made in a time when actually figuring them out was very complicated to me. It took me a very long time to understand them, so it is somewhat excusable that I did not want to mess with them when it took me such a long time to get them right.

Right now they are just another standard feature that is fully understod by me.

Each of the join points is different and no doubt, I will fiddling with the cuda one before long, so the way they are now is the correct way of handling them.