I will deal with it tomorrow.
Improving .NET interop again seems like it would take a few hours. Screw that. I am calling it day here.

What I will do is debug, deal with this IsStatic and then start putting in basic extern ops and maybe thanks to that, lighten up the language a little.

After that I will able to start work on the Cuda backend in a more substantial fashion. I've yet to put in union types on its end.