Let me start with this.
Test2 stackoverflows. Yeah, this is going to be a slog.

Also, I've decided on a change in policy. I am going to remove load_assembly from the language and have them passed in at the start of compilation.

This is what I am going to do anyway to pass includes into the Cuda segment so it makes sense to do it like this.

Now let me figure out what is wrong.