4:30pm. I've been really deep in thought for the past few hours. I understand how to do both Extern and the .NET interop, but I am not really happy how things are going to go. Here are my conclussions:
1) Regarding Extern. I've been extensively debating whether I should compile everything to Extern operations in Spiral, but then I realized what that would trully mean.

First it would mean the compiler going through 3.3k bindings on each compile just mscorlib.

But as bad as that is, the reality is much worse. Spiral has that map filtering thing which it needs in order to ensure that join points due no diverge, that means that due to repeated pointless filtering of all the variables in scope, what would happen is more like it would go 3.3k^2/2 the number of bindings. That would be really, really bad.

So Extern is out for .NET interop. There are some more problems.

Given Spiral's functional nature I would not be able to put the methods and static method directly into the type, rather I would have to separate them and explicitly call upon them. Yeah, I know that Spiral can do OO, but I do not see the way to compile it otherwise given how types can be recursive in .NET. It would be different if I was writting things by hand, but forget it. I can't see far enough into it just from here.

There are just too many hurdles to attempt .NET interop using Extern operations at the current juncture.

Spiral is not the pinnacle programming languages done by a superhuman being, it is the crystalization of my desire, weakness and courage.

I think the true language will defintely be a hugely efficient language of the likes I can hardly even comprehend. It will be a functional language because first-class functions are such an integral element of reality. Even if programming does not scale, I think that throwing 10^30 times my brainpower at the problem should push the limits quite a ways.

That brings me to second point.

2) The real deciding factor is simply the inefficieny of Extern which is an offshot of the inefficiency of Spiral at compile time.

So what I am going to do is convert the `System.Type` to an inbuilt type. I should have done this ages ago, but I was too intimidated at the prospect of doing my own substitution and was stubborn in looking for alternatives to doing it on my own.

It won't be too difficult and it will give me complete composability without the disadvantages of going through Extern.

To be honest, I do not feel like doing right now though. I'll guess I'll try and start something at least.