I am ready to start for the day.
Right now I should be working on .NET interop, but yesterday's performance exercise brought some inspiration to me today.

I realized that if I want to make the language scale, I need to stage the hashconsing and the renaming for as long as possible, all the way into the join points. At that point, I will be able to attain the optimal performance with the language from an algorithmic pespective, but doing this would raise the complexity of the language once again.

When I have a new idea I am usually raring to go, but now I am really considering whether I should just leave it on the backburner for the time being.

Yeah, let me just focus on .NET interop for now. It is not like I can't do this at a later point when I am not swamped with work.

Let me deal with .NET introp for now. It will take me a day or two now that I know how to do it and I will be able to test out the idea above after that.