Right now my fatigue is over the limit and I forgot to put the generic type arguments in the class record. I'll deal with it tomorrow, right now I honestly feel sleepy.
I think I am not about a third of the way in in terms of overall completeness, and maybe 70% there in terms of implementation. I am going to have to put in HM inference in order to do variable substitution after I deal with the first testing pass.

But after that, for a few years, I will not have to touch .NET interop stuff again. I will be able to make 90% use of the .NET libraries (because of the lack of subtyping) and that will be enough to get me anywhere I need.

After that I will put the final screw on performance and hopefully finally dive into the damn backend.

Every single time I've said that I am done with the language I've been proven wrong, but eventually it will yield to my effort.

I won't worry about finishing it before the time for the next update. Still, two months of full time work is plenty of time to make something good, so maybe I will get lucky?