"6:40pm. The example typechecks. let (var_15: int64) = System.Int64.MinValue
Here is the static method, printed on line 149. I think with this I have enough to call it a day. I will finish the rest of the tests tomorrow.

...But actually, let me try running them anyway.

I want to see if it will work.

I am not too tired.

6:40pm. It took 22s to finish all the tests. The thing really is a lot slower right now since I added the latest changes.

Aggghhh...piece of shit .NET interop.

...Also I hate it when I get diffs for everything even though nothing changed.
let test33 = // 0.42s
    "test33",[],"Does a simple loop have superlinear scaling?",
    """
inl rec loop = function
    | i when i > 0 -> loop (i-1)
    | 0 -> ()
loop 50000
    """
Wow, this now literally takes 5.2 seconds. What the fuck?

6:45pm. Let me figure out what is wrong with Cuda. It says that it cannot find IsNull? Ah, yes.
    inl print_to_standard_output = closure_of (inl _ args -> args.get_Data() |> writeline)
    inl add_handler event =
        system .System.Diagnostics.DataReceivedEventHandler print_to_standard_output
        |> event_add_handler process event

I fixed the isNull thing. Now comes this.

6:55pm.
    inl print_to_standard_output = term_cast (inl args -> args.get_Data() |> writeline) (system .System.Diagnostics.DataReceivedEventArgs)

    process.ErrorDataReceived.Add print_to_standard_output
//    process.OutputDataReceived.Add print_to_standard_output
Let me do it like this.

Now let me try it again.

7pm.
Time for prepass: 00:00:00.0749289
Cannot find static method get_Data in the .NET type.
Error trace on line: 2, column: 1 in file "Cuda".

Ugh...
    inl print_to_standard_output =
        term_cast (inl args -> args.get_Data() |> writeline)
            (system .System.Diagnostics.DataReceivedEventArgs)
Shit. What do I do here?
TyV (0, DotNetTypeT <tag 2142>)
Cannot find static method get_Data in the .NET type.

Well, it is TyV when it comes in.

So it should not be looking for static methods.

Let me give something a shot.

7pm.
let f t t' = t |> tyt, t'

Instead of TyT it here, I should give it a TyV even I have fake it.

...No.

No.

7:05pm. No, shit. No.

I've coded myself into a hole. The automatic conversions and the staging of standard types is a decisive failure.

What I need to do is get rid of the delay in the compilation of types with no type arguments and forget about using static methods on primitive types.

They are not important, not as much as the rest anyway.

Seriously, fuck fuck .NET. Fuck it, fuck it. Fuck OO. Fuck it with a rake.

Well, given the sheer slowness of it now, I doubt that compiling the standard types right away will cause it to run much slower.

I am going to have to get to the bottom of just what the hell is happening here."

I will stop here. There is no point in going any further than this today.

.NET is really proving much harder to do interop with than one would expect. I do not understand either why it got 2x slower again.

I hate this, really. By this, I mean .NET. But I did say that perfect interop is not important so I will not let it bother me.

Tomorrow I will simplify the implementation even if it makes it slower and then do performance debugging again. I'll add some of my recent ideas to the mix.