Here is the experimentation that I did. With this I have enough to establish equality between type definitions. With this I can use them in dictionaries and hence have enough to actually do the rewrite.
One thing I'd like to know is how the assemblies are loaded from the global namespace, I think it was called MAC or something. Well, figuring that out won't be too hard, it will be an easy question for SO.

The important thing is that the rewrite now that I have all the pieces won't be hard to do at all.

I'll do it all tomorrow. Today, I am too tired.

It is time to call it a day.