I did some design improvement for the interop functions. I might have fixed some bugs that I was not even aware of. Also I refactored the memoize function a bit.
It feel like I did not accomplish much today.

My preference would be to make progress every day, but .NET is really trying its best to make it as hard as possible to interop with it.

If I ever get truck'd to the past, I'd probably try implementing Spiral in Racket instead or something.

As it is though, .NET really is possibly the #1 or #2 choice for making a functional language regardless of the difficulties it is giving me right now.

Well, just forget about it. Last time I asked Jb a question it did take him half a week to get back to me, so I won't just sit on my laurels waiting for him to get to me.

I will go forward with the assumption that Cecil won't pan out.

What I will do now is push the renaming and consing all the way into the renamer.

This is purely a speed optimization as all the filtering the pevaller now does leads it to being a polynomial time for the number of bindings in use.

I need to cut that before the 0.1 release.

Excepting the Cecil rewrite of the interop module, what I will do whne I am done with this is get back to work on the Cuda backend.

I am going to remove Cuda constants directly from the language and Extern them instead.

As the last thing for this commit, let me just say that today's simplification of the introp module will also help the Cecil rewrite when it comes it.

I did really well to figure it out.

Well, let me do a little bit to at least get the renaming thing going.