I have no idea how to do the prepass. Thanks to it, the net speed improvement is only 4x instead 30x on that particular case. Balls.
Well, it was not that important.

I've done enough for the day.

In the past week I've implemented full .NET introp and now this.

At this point the language cannot be considered anything other than complete. At some point it will have enough to be considered that.

Now that I have all these significant pieces I can look forward into getting back to work on the Cuda backend.

What was I even doing before this latest diversion? The Cuda Array map?

Maybe along the way I will change the ints to be 32-bit by defualt. I am really not getting much gains from having them be 64-bit.

...

Ah, yes, I remember now what I was aiming at.

It is true that the arrays functions are a goal, but the first thing I will do tomorrow is replace the CudaIdx and those built in constants and syncthreads with Extern calls.

The problem with the array function is that I am still yet to do some preliminaries.

Apart from adding Extern ops on the Cuda side, what I still need to do is making sure that all the variables that pass on the Cuda side are valid.

And apart from that I need to convert the Cuda device memory pointers to Cuda arrays.

I'll need to add a special op for that.

I think this about covers it. I will stop here as doing too much in a day has the effect of raising my stress too high.