I did some work in the other repo and the rewrite will be done if I can deal with the issue there, but now let me get back on track here. I'll forget about the thing for the time being and forget it even exists.
Right now I am distracted by the work being done in my room and am trying to gather my focus.

The Cecil rewrite also made me go off track.

"It is true that the arrays functions are a goal, but the first thing I will do tomorrow is replace the CudaIdx and those built in constants and syncthreads with Extern calls.

The problem with the array function is that I am still yet to do some preliminaries.

Apart from adding Extern ops on the Cuda side, what I still need to do is making sure that all the variables that pass on the Cuda side are valid.

And apart from that I need to convert the Cuda device memory pointers to Cuda arrays.

I'll need to add a special op for that."

The above is what I wrote yesterday,

Let me put in some Cuda externs, replace the constants and then I will move to typechecking the arguments that pass the Cuda join point.

Also while I am at it, why don't I get rid of the Cuda kernels op and replace it with an extern op?