Today between 1:40pm and 5pm I had to be away from the coputer so I could not do as much as I intended, but nevertheless I did quite a bit in the time that was alloted to me.
-Managed to complete the Mono.Cecil rewrite that will go unused until I can get it to work completely.
-Removed a bunch of useless constants and replaced them with Extern ops.
-Added type checks for variables on the Cuda join point.
-Rewrote the two codegens so that trace is done using exceptions. This saves me from having to take apart the code in order to pass the parameter along.
-Added support for char on the Cuda side.
-Changed the type of BoolT to be "char" on Cuda side.
-Added support for CharT on the Cuda side.
-Added the coercion op for CudaGlobalArray.
-Extended the CudaGlobalArray to take a token type.
-Added the `Art` prefix to all the array types.

It is not much as it sounds, there just a lot of different small things.

With this the stage is set. What I will get to work on starting tomorrow is a proper Cuda array type. CudaDeviceVariable is almost enough, but the problem with it is that it is not a struct, but a reference type with the size included. For the Cuda array type to be directly passable, it needs to be a raw pointer.

As hard as it sound to believe, once I deal with that I will have everything I need to begin work on the Spiral ML library.

I am nearly there.

I keep saying that I've been working on the language since February, but if I count where I threw in the towel on implementing layer normalization due to lack of appropriate abstraction, then it would be over a year by now.

I can't believe just how much time doing the right thing cost me.

So far sticking to my principles has brought me only pain; the programming devil is making me pay through my nose in upfront costs.

I think that from here I might starting seeing some actual returns on my investment.

It has been really a long time in coming.