I did some redesigning of ArrayN. I renamed it to HostTensor and adjusted the examples. In test83 I even found an error with the original that I thought I had fixed.
Let me call it a day here. I will start working on the new tensor modules with the tuple of arrays layout tomorrow. It is not the kind of thing I can do light lightly and it will take me a day or three to get it done for both host and Cuda.

But it should not be too difficult.

The harded thing to decide for me right now is how eactly should I lead them.

...I guess I will start with the TensorHelpers then.