I think I've settled on the design.
    inl map {map_op ins outs} = cuda
        inl from = blockIdx.x * blockDim.x + threadIdx.x
        inl near_to = total_size ins
        inl by = gridDim.x * blockDim.x
        for {from near_to by body = inl {i} ->
            set ins i (map_op (index outs i))
            }

What I will do is in fact let the map kernel be like  this, but I will not aim to make the CudaTensor be the same on the host and inside the kernel. Rather I will transform it to the CudaKernelTensor where the ar will be a Cuda array instead of the ManagedCuda.CudaDeviceVariable. Everything else will be the same and I will be able to resuse code between the two types.

For the past few hours, I've been focusing on how to increase the composability of the code, but really, I do not think I will get what I want on the kernel level. As time goes on the GPU will get more generalist capability, but that time is not here yet.

What I should focus on with Spiral is on eroding module boundaries and turning that into power.

So then it should be fine to add one extra tensor, but otherwise go with the above plan. This is without a doubt the correct conclussion.