With this, the new map kernel is officially complete.
Awesome, this just too awesome.

I've noticed that I need to make some changes to the way I do testing.

My pace is simply too slow which also leads to me losing focus.

Deep long contemplation has really served me well up to now. Up to now, my main enemies were compiler bugs and I've had to continually add features to the language. But as yesterday showed, at this point I would be better served simply trusting the compiler that I worked so long on. It is definitely getting more stable and I can't recall the last time I had a bug in it.

What I will be better served with would be to pick up the pace and not be so afraid of getting a error.

I've been conditioned to be afraid of them, because for a long time, an error often meant having to spend hours hacking on the compiler in addition to having to think whether the program was correct in the first place, but now, the program in fact does what I think it does.

That has been a trend.

Yesterday I've considered compiling the language into a .dll for the first time so I do not have to compile it together with the examples.

I'll keep doing it like this for a few more days and if I see no need to keep hacking the language, I will change the process.

Now...what comes next?

At this point I have everything I need to port the old Spiral ML library to Spiral Language.

Just being able to say that last line is pleasing to me. What will I feel once I actually finish the port and do what I could not do in October of last year which is dealing with layer normalization?

That thing brought me utterly to my knees before.

I remember how while I was struggling with how to abstract the kernel back in November how some guy just came to the ML sub and posted a Python and C++ implementation.

It did not shake me in the slightest and neither did it affect my resolve. And the end of result of that drive will be this language.

Well, since I've bene at it for a few hours, let me stop here for a while.

I need to think what I will do next, but I am leaning towards starting the port. The is no need to delay this any longer and there are no more obstacles.