"7pm. I can't stop reading this.
http://lambda-the-ultimate.org/node/220#comment-1695

"Also, I think that in twenty years there will no longer be any debate about static vs. dynamic typing, and that instead most languages will instead provide a smooth continuum between the two, as people realize that there is no fundamental dichotomy there. Also languages will have more than one level of types: the types themselves will have types, and so on. But we will be able to treat types in a similar way to the way in which we treat values: we will be doing computations with them. "

"I wonder if one of the "levels of types" in this picture might be latent types? ;o)
Another comment of Frank's which is right in line with what I think is "I have a feeling that DT languages are not at the 'end of the line' and could go further by adopting online partial evaluation techniques"."

The above is from 2004.

7:10pm. Latent type. This is the first I've heard of this term. Does Spiral have that?

7:15pm. http://lambda-the-ultimate.org/classic/message6475.html#6506

"There will be an increased emphasis on efficiency in the future, and I think we will see fewer aggressively optimizing compilers; instead it will be the programmer's job to write his programs in such a way that they are pretty much guaranteed to be efficient with any reasonable compiler. This sounds like a step backward, but it won't be because we will be better able to separate concerns of specification from implementation. "

It is like this guy is pretty much predicting the emergence of Spiral.

I should have not gone into trading instead of programming.

http://www.di.ens.fr/~cousot/COUSOTpapers/publications.www/Cousot-POPL97-p316-331-1997.pdf
Types as Abstract Interpretations

7:25pm. Ok, enough I am going to die if I continue. I will save the links here.

http://www.imm.dtu.dk/~hrni/PPA/ppa.html
http://lambda-the-ultimate.org/node/220#comment-1695
http://lambda-the-ultimate.org/classic/message6475.html#6506

https://pdfs.semanticscholar.org/7098/d121bf9d749d67fb902d138973e8f09b516a.pdf
Building Complete Abstract Interpretations in a Linear Logic-based Setting"

I finished a while ago and am now just chilling.

I've changed my mind about starting tomorrow, I think I will defer it for one more day. I want to spend some time studying abstract interpretation tomorrow.

Then I will get to work on the allocator.

Once I am done with that I will make sure to put in union types on the Cuda side. What I have right now is just annoying me.